requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
PyYAML
gitpython
google-generativeai